+++ iceoryx_posh/include/iceoryx_posh/internal/mepoo/mem_pool.hpp
@@ -37,12 +37,12 @@ struct MemPoolInfo
     MemPoolInfo(const uint32_t usedChunks,
                 const uint32_t minFreeChunks,
                 const uint32_t numChunks,
-                const uint32_t chunkSize) noexcept;
+                const uint64_t chunkSize) noexcept;
 
     uint32_t m_usedChunks{0};
     uint32_t m_minFreeChunks{0};
     uint32_t m_numChunks{0};
-    uint32_t m_chunkSize{0};
+    uint64_t m_chunkSize{0};
 };
 
 class MemPool
@@ -51,7 +51,7 @@ class MemPool
     using freeList_t = concurrent::MpmcLoFFLi;
     static constexpr uint64_t CHUNK_MEMORY_ALIGNMENT = 8U; // default alignment for 64 bit
 
-    MemPool(const greater_or_equal<uint32_t, CHUNK_MEMORY_ALIGNMENT> chunkSize,
+    MemPool(const greater_or_equal<uint64_t, CHUNK_MEMORY_ALIGNMENT> chunkSize,
             const greater_or_equal<uint32_t, 1> numberOfChunks,
             iox::BumpAllocator& managementAllocator,
             iox::BumpAllocator& chunkMemoryAllocator) noexcept;
@@ -62,7 +62,7 @@ class MemPool
     MemPool& operator=(MemPool&&) = delete;
 
     void* getChunk() noexcept;
-    uint32_t getChunkSize() const noexcept;
+    uint64_t getChunkSize() const noexcept;
     uint32_t getChunkCount() const noexcept;
     uint32_t getUsedChunks() const noexcept;
     uint32_t getMinFree() const noexcept;
@@ -75,7 +75,7 @@ class MemPool
     /// @param[in] chunkSize is the size of the chunk
     /// @param[in] rawMemoryBase it the pointer to the raw memory of the MemPool
     /// @return the pointer to the chunk
-    static void* indexToPointer(const uint32_t index, const uint32_t chunkSize, void* const rawMemoryBase) noexcept;
+    static void* indexToPointer(const uint32_t index, const uint64_t chunkSize, void* const rawMemoryBase) noexcept;
 
     /// @brief Converts a pointer to a chunk in the MemPool to an index
     /// @param[in] chunk is the pointer to the chunk
@@ -83,15 +83,15 @@ class MemPool
     /// @param[in] rawMemoryBase it the pointer to the raw memory of the MemPool
     /// @return the index to the chunk
     static uint32_t
-    pointerToIndex(const void* const chunk, const uint32_t chunkSize, const void* const rawMemoryBase) noexcept;
+    pointerToIndex(const void* const chunk, const uint64_t chunkSize, const void* const rawMemoryBase) noexcept;
 
   private:
     void adjustMinFree() noexcept;
-    bool isMultipleOfAlignment(const uint32_t value) const noexcept;
+    bool isMultipleOfAlignment(const uint64_t value) const noexcept;
 
     RelativePointer<void> m_rawMemory;
 
-    uint32_t m_chunkSize{0U};
+    uint64_t m_chunkSize{0U};
     /// needs to be 32 bit since loffli supports only 32 bit numbers
     /// (cas is only 64 bit and we need the other 32 bit for the aba counter)
     uint32_t m_numberOfChunks{0U};
