+++ iceoryx_posh/test/moduletests/test_popo_used_chunk_list.cpp
@@ -34,7 +34,7 @@ class UsedChunkList_test : public Test
     void SetUp() override
     {
         static constexpr uint32_t NUM_CHUNKS_IN_POOL = 100U;
-        static constexpr uint32_t CHUNK_SIZE = 128U;
+        static constexpr uint64_t CHUNK_SIZE = 128U;
         MePooConfig mempoolconf;
         mempoolconf.addMemPool({CHUNK_SIZE, NUM_CHUNKS_IN_POOL});
 
@@ -46,7 +46,7 @@ class UsedChunkList_test : public Test
 
     SharedChunk getChunkFromMemoryManager()
     {
-        constexpr uint32_t USER_PAYLOAD_SIZE{32U};
+        constexpr uint64_t USER_PAYLOAD_SIZE{32U};
         auto chunkSettingsResult =
             iox::mepoo::ChunkSettings::create(USER_PAYLOAD_SIZE, iox::CHUNK_DEFAULT_USER_PAYLOAD_ALIGNMENT);
         IOX_ENSURES(chunkSettingsResult.has_value());
